Remove reprocessed, fixed exposures from bad_expid
Frank's reprocessing of DECam fixed a number of images with bugs in their processing.  These may now be removed from DR9.

I left 368186 in the bad_expid list because it is clear that there are significant scattering halos around the bright stars, so there are probably clouds.  721209, 736317, 780010, and 420479 also remain in the list, because we either didn't reprocess them, they had weird propid's that we wouldn't include anyway, or in the case of 420479, it looks like I accidentally transcribed that from the 'correct' 420749.   That's from a propid we no longer include.