Add additional bad focus images to bad_expid file
Add some images to the bad expid file for 90prime where 90prime showed clear larger-than-typical residuals.  These look to be out of focus / cometary / donut-y.  However, the earlier images on this night would probably be recoverable with a better PSF analysis; they're not that cometary.  But ~0.1 mag residuals in the photometry is not really okay.