Modified runbrick-gpu.sh to add command line options: -ngpu N -gpu_ids 0,1,2,3 -threads_per_gpu N
runbrick.py modified to support mix of GPU and CPU threads and
multi-gpu.  threads arg gives total threads (eg 128) and threads_per_gpu
* ngpu gives the number of GPU threads.  GPU threads start first in
round robin, so that blob 1 goes to GPU 1, blob 2 to GPU 2, ...,
blob 5 to GPU 1, in a 4 GPU setup and all GPU threads are filled first
before CPU threads so that the larger blobs are started implicitly on
GPUs.